refactor: flatten checkpoint/k8s subpackage and reorganize checkpoint modules
- Flatten checkpoint/k8s/ subpackage: OCI types/getters → oci.go,
  K8s+containerd types (DiscoveryClient, DetectVolumeTypeFromPath) → k8s.go
- Delete pkg/config/ and pkg/common/metadata.go, consolidate into
  checkpoint package and cmd/agent/config.go
- Split checkpoint.go: Config → config.go, Save/Load I/O → storage.go,
  filesystem types → filesystem.go
- Each module owns its metadata type and New*() constructor:
  NewK8sMetadata, NewFilesystemMetadata, NewMountMetadata, NewNamespaceMetadata
- NewCheckpointMetadata is now a trivial assembly of per-module builders
- Remove one-liner wrappers (GetCheckpointDir, OpenImageDir,
  AddExternalNamespace, etc.) — inline at call sites
- Delete unused functions (NeedsCRIUConfFile, GetNamespaceKey,
  FormatExternalNamespace, IsNetNamespaceExternal, GetHostname, etc.)
- Remove redundant restore-side tar excludes (already handled at
  checkpoint time via OCI bind mount detection and values.yaml exclusions)
