+++ deploy/chrek/pkg/checkpoint/checkpoint.go
@@ -9,15 +9,46 @@ import (
 	"time"
 
 	criu "github.com/checkpoint-restore/go-criu/v7"
+	criurpc "github.com/checkpoint-restore/go-criu/v7/rpc"
 	"github.com/sirupsen/logrus"
 	"google.golang.org/protobuf/proto"
 
-	checkpointk8s "github.com/ai-dynamo/dynamo/deploy/chrek/pkg/checkpoint/k8s"
-	"github.com/ai-dynamo/dynamo/deploy/chrek/pkg/config"
+	"github.com/ai-dynamo/dynamo/deploy/chrek/pkg/common"
 )
 
+// CheckpointMetadata is saved as metadata.yaml at checkpoint time, loaded at restore.
+type CheckpointMetadata struct {
+	CheckpointID string    `yaml:"checkpointId"`
+	CreatedAt    time.Time `yaml:"createdAt"`
+
+	CRIU       CRIUConfig          `yaml:"criu"`
+	K8s        K8sMetadata         `yaml:"k8s"`
+	Filesystem FilesystemMetadata  `yaml:"filesystem"`
+	Mounts     []MountMetadata     `yaml:"mounts"`
+	Namespaces []NamespaceMetadata `yaml:"namespaces"`
+}
+
+// NewCheckpointMetadata assembles a CheckpointMetadata from per-module builders.
+func NewCheckpointMetadata(
+	checkpointID string,
+	criuCfg CRIUConfig,
+	k8s K8sMetadata,
+	filesystem FilesystemMetadata,
+	mounts []MountMetadata,
+	namespaces []NamespaceMetadata,
+) *CheckpointMetadata {
+	return &CheckpointMetadata{
+		CheckpointID: checkpointID,
+		CreatedAt:    time.Now().UTC(),
+		CRIU:         criuCfg,
+		K8s:          k8s,
+		Filesystem:   filesystem,
+		Mounts:       mounts,
+		Namespaces:   namespaces,
+	}
+}
+
 // CheckpointParams holds per-checkpoint identifiers for a checkpoint operation.
-// This is separate from config.CheckpointConfig which holds static CRIU settings.
 type CheckpointParams struct {
 	ContainerID   string
 	ContainerName string // K8s container name (for K8s API volume type lookup)
@@ -32,175 +63,64 @@ type CheckpointParams struct {
 type Result struct {
 	CheckpointID  string
 	CheckpointDir string
-	Data          *config.CheckpointData
+	Data          *CheckpointMetadata
+}
+
+// containerState holds all introspected container state needed for checkpoint.
+type containerState struct {
+	ContainerInfo *ContainerInfo
+	PID           int
+	RootFS        string
+	UpperDir      string
+	Mounts        []MountMapping
+	Namespaces    map[NamespaceType]*NamespaceInfo
 }
 
 // Checkpointer performs CRIU checkpoint operations
 type Checkpointer struct {
-	discoveryClient *checkpointk8s.DiscoveryClient
-	k8sClient       *checkpointk8s.K8sClient // Optional: for accurate volume type discovery from K8s API
+	discoveryClient *DiscoveryClient
 	log             *logrus.Entry
 }
 
 // NewCheckpointer creates a new checkpointer
-func NewCheckpointer(discoveryClient *checkpointk8s.DiscoveryClient) *Checkpointer {
+func NewCheckpointer(discoveryClient *DiscoveryClient) *Checkpointer {
 	return &Checkpointer{
 		discoveryClient: discoveryClient,
 		log:             logrus.WithField("component", "checkpointer"),
 	}
 }
 
-// WithK8sClient sets an optional Kubernetes client for accurate volume type discovery.
-// When set, volume types are fetched from the K8s API instead of being inferred from paths.
-func (c *Checkpointer) WithK8sClient(client *checkpointk8s.K8sClient) *Checkpointer {
-	c.k8sClient = client
-	return c
-}
-
-// Checkpoint performs a CRIU dump of a container
-func (c *Checkpointer) Checkpoint(ctx context.Context, params CheckpointParams, cfg *config.CheckpointConfig) (*Result, error) {
+// Checkpoint performs a CRIU dump of a container.
+// The operation has three phases: introspect, configure, capture.
+func (c *Checkpointer) Checkpoint(ctx context.Context, params CheckpointParams, cfg *Config) (*Result, error) {
 	if cfg == nil {
 		return nil, fmt.Errorf("checkpoint config is required")
 	}
 	checkpointStart := time.Now()
 	c.log.Info("=== Starting checkpoint operation ===")
 
-	// 1. Resolve container to get PID
-	resolveStart := time.Now()
-	containerInfo, err := c.discoveryClient.ResolveContainer(ctx, params.ContainerID)
-	if err != nil {
-		return nil, fmt.Errorf("failed to resolve container: %w", err)
-	}
-	pid := int(containerInfo.PID)
-	c.log.WithField("duration", time.Since(resolveStart)).Info("Container resolution completed")
-
-	// 2. Create checkpoint directory
-	checkpointDir := config.GetCheckpointDir(params.CheckpointDir, params.CheckpointID)
+	checkpointDir := filepath.Join(params.CheckpointDir, params.CheckpointID)
 	if err := os.MkdirAll(checkpointDir, 0700); err != nil {
 		return nil, fmt.Errorf("failed to create checkpoint directory: %w", err)
 	}
 
-	// 3. Introspect container state
-	introspectStart := time.Now()
-	rootFS, err := GetRootFS(pid)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get rootfs: %w", err)
-	}
-	mounts, err := GetKubernetesVolumeMounts(pid)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get mounts: %w", err)
-	}
-	namespaces, err := GetAllNamespaces(pid)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get namespaces: %w", err)
-	}
-	c.log.WithField("duration", time.Since(introspectStart)).Info("Container introspection completed")
-
-	// 4. Open image directory FD
-	imageDir, imageDirFD, err := OpenImageDir(checkpointDir)
+	// Phase 1: Introspect container state
+	state, err := c.introspect(ctx, params.ContainerID)
 	if err != nil {
 		return nil, err
 	}
-	defer imageDir.Close()
 
-	// 5. Build CRIU options from config
-	criuParams := CRIUDumpParams{
-		PID:        pid,
-		ImageDirFD: imageDirFD,
-		RootFS:     rootFS,
-	}
-	criuOpts := BuildCRIUOpts(&cfg.CRIU, criuParams)
-
-	// 6. Create CRIU config file if needed (for options not available via RPC)
-	if cfg.CRIU.NeedsCRIUConfFile() {
-		if err := cfg.CRIU.Validate(); err != nil {
-			return nil, err
-		}
-		configPath := filepath.Join(checkpointDir, config.CheckpointCRIUConfFilename)
-		configContent := cfg.CRIU.GenerateCRIUConfContent()
-		if err := os.WriteFile(configPath, []byte(configContent), 0644); err != nil {
-			return nil, fmt.Errorf("failed to write CRIU config file: %w", err)
-		}
-		criuOpts.ConfigFile = proto.String(configPath)
-		c.log.WithFields(logrus.Fields{
-			"config_path": configPath,
-			"lib_dir":     cfg.CRIU.LibDir,
-		}).Info("Created CRIU config file")
-	}
-
-	// 7. Configure external mounts and namespaces
-	if err := ConfigureExternalMounts(criuOpts, pid, containerInfo); err != nil {
+	// Phase 2: Configure CRIU options and build checkpoint metadata
+	criuOpts, data, err := c.configure(state, params, cfg, checkpointDir)
+	if err != nil {
 		return nil, err
 	}
-	netNsInode := ConfigureExternalNamespaces(criuOpts, namespaces, cfg.CRIU.ExternalMounts)
-	if netNsInode > 0 {
-		c.log.WithField("inode", netNsInode).Debug("Marked network namespace as external")
-	}
-	for _, extMount := range cfg.CRIU.ExternalMounts {
-		c.log.WithField("external", extMount).Debug("Added external mount mapping")
-	}
 
-	// 7b. Configure mounts to skip from configured prefixes (for cross-node restore)
-	skipMounts, err := ConfigureSkipMounts(criuOpts, pid, cfg.SkipMountPrefixes)
+	// Phase 3: Capture — CRIU dump, /dev/shm, rootfs diff
+	criuDumpDuration, err := c.capture(criuOpts, data, state, checkpointDir)
 	if err != nil {
-		c.log.WithError(err).Warn("Failed to configure skip mounts")
-	} else if len(skipMounts) > 0 {
-		c.log.WithFields(logrus.Fields{
-			"prefixes":    cfg.SkipMountPrefixes,
-			"skip_mounts": skipMounts,
-			"count":       len(skipMounts),
-		}).Info("Configured mounts to skip for cross-node restore")
+		return nil, err
 	}
-
-	// 8. Get overlay upperdir for rootfs diff capture
-	upperDir, upperDirErr := GetOverlayUpperDir(pid)
-	if upperDirErr != nil {
-		c.log.WithError(upperDirErr).Warn("Could not get overlay upperdir - rootfs diff will not be captured")
-	} else {
-		c.log.WithField("upperdir", upperDir).Debug("Found overlay upperdir")
-	}
-
-	// 9. Build checkpoint data with config and container state
-	metaCfg := CheckpointDataBuilderConfig{
-		CheckpointID:  params.CheckpointID,
-		NodeName:      params.NodeName,
-		ContainerID:   params.ContainerID,
-		ContainerName: params.ContainerName,
-		PodName:       params.PodName,
-		PodNamespace:  params.PodNamespace,
-		PID:           pid,
-	}
-	data := BuildCheckpointData(ctx, metaCfg, cfg, containerInfo, mounts, namespaces, c.k8sClient, c.log)
-	if upperDir != "" {
-		data.UpperDir = upperDir
-	}
-
-	if err := config.SaveCheckpointData(checkpointDir, data); err != nil {
-		return nil, fmt.Errorf("failed to save checkpoint data: %w", err)
-	}
-
-	// 10. Execute CRIU dump via go-criu
-	criuDumpStart := time.Now()
-	criuClient := criu.MakeCriu()
-	if err := criuClient.Dump(criuOpts, nil); err != nil {
-		c.log.WithField("duration", time.Since(criuDumpStart)).Error("CRIU dump failed")
-		return nil, fmt.Errorf("CRIU dump failed: %w", err)
-	}
-	criuDumpDuration := time.Since(criuDumpStart)
-	c.log.WithField("duration", criuDumpDuration).Info("CRIU dump completed successfully")
-
-	// 11. Capture /dev/shm contents
-	// This must happen after CRIU dump since we want the final process state
-	shmCaptureStart := time.Now()
-	if err := CaptureDevShm(pid, checkpointDir, c.log); err != nil {
-		c.log.WithError(err).Warn("Failed to capture /dev/shm contents")
-	}
-	c.log.WithField("duration", time.Since(shmCaptureStart)).Info("/dev/shm capture completed")
-
-	// 12. Capture rootfs diff and deleted files
-	rootfsCaptureStart := time.Now()
-	CaptureRootfsState(upperDir, checkpointDir, data, c.log)
-	c.log.WithField("duration", time.Since(rootfsCaptureStart)).Info("Rootfs capture completed")
 
 	totalDuration := time.Since(checkpointStart)
 	c.log.WithFields(logrus.Fields{
@@ -213,4 +133,125 @@ func (c *Checkpointer) Checkpoint(ctx context.Context, params CheckpointParams, 
 		CheckpointDir: checkpointDir,
 		Data:          data,
 	}, nil
+}
+
+// introspect resolves the container and gathers all runtime state from containerd and /proc.
+func (c *Checkpointer) introspect(ctx context.Context, containerID string) (*containerState, error) {
+	containerInfo, err := c.discoveryClient.ResolveContainer(ctx, containerID)
+	if err != nil {
+		return nil, fmt.Errorf("failed to resolve container: %w", err)
+	}
+	pid := int(containerInfo.PID)
+
+	rootFS, err := GetRootFS(pid)
+	if err != nil {
+		return nil, fmt.Errorf("failed to get rootfs: %w", err)
+	}
+	upperDir, err := GetOverlayUpperDir(pid)
+	if err != nil {
+		return nil, fmt.Errorf("failed to get overlay upperdir: %w", err)
+	}
+	mounts, err := GetKubernetesVolumeMounts(pid)
+	if err != nil {
+		return nil, fmt.Errorf("failed to get mounts: %w", err)
+	}
+	namespaces, err := GetAllNamespaces(pid)
+	if err != nil {
+		return nil, fmt.Errorf("failed to get namespaces: %w", err)
+	}
+
+	return &containerState{
+		ContainerInfo: containerInfo,
+		PID:           pid,
+		RootFS:        rootFS,
+		UpperDir:      upperDir,
+		Mounts:        mounts,
+		Namespaces:    namespaces,
+	}, nil
+}
+
+// configure builds CRIU options and checkpoint metadata from introspected state and config.
+func (c *Checkpointer) configure(
+	state *containerState,
+	params CheckpointParams,
+	cfg *Config,
+	checkpointDir string,
+) (*criurpc.CriuOpts, *CheckpointMetadata, error) {
+	// Open image directory FD
+	imageDir, imageDirFD, err := common.OpenPathForCRIU(checkpointDir)
+	if err != nil {
+		return nil, nil, fmt.Errorf("failed to open image directory: %w", err)
+	}
+	defer imageDir.Close()
+
+	// Build CRIU options from config
+	criuOpts := BuildCRIUOpts(&cfg.CRIU, CRIUDumpParams{
+		PID:        state.PID,
+		ImageDirFD: imageDirFD,
+		RootFS:     state.RootFS,
+	})
+
+	// Write CRIU config file (for options unavailable via RPC)
+	configPath := filepath.Join(checkpointDir, CheckpointCRIUConfFilename)
+	if err := os.WriteFile(configPath, []byte(cfg.CRIU.GenerateCRIUConfContent()), 0644); err != nil {
+		return nil, nil, fmt.Errorf("failed to write CRIU config file: %w", err)
+	}
+	criuOpts.ConfigFile = proto.String(configPath)
+
+	// Configure external mounts and namespaces
+	if err := ConfigureExternalMounts(criuOpts, state.PID, state.ContainerInfo); err != nil {
+		return nil, nil, err
+	}
+	ConfigureExternalNamespaces(criuOpts, state.Namespaces, cfg.CRIU.ExternalMounts)
+
+	// Configure mounts to skip (for cross-node restore)
+	if _, err := ConfigureSkipMounts(criuOpts, state.PID, cfg.SkipMountPrefixes); err != nil {
+		return nil, nil, err
+	}
+
+	// Build and save checkpoint metadata
+	oci := extractOCIState(state.ContainerInfo)
+	metadata := NewCheckpointMetadata(
+		params.CheckpointID,
+		cfg.CRIU,
+		NewK8sMetadata(params, state.PID, state.ContainerInfo),
+		NewFilesystemMetadata(cfg.RootfsExclusions, state.UpperDir, oci),
+		NewMountMetadata(state.Mounts, oci),
+		NewNamespaceMetadata(state.Namespaces),
+	)
+
+	if err := SaveCheckpointMetadata(checkpointDir, metadata); err != nil {
+		return nil, nil, fmt.Errorf("failed to save checkpoint metadata: %w", err)
+	}
+
+	return criuOpts, metadata, nil
+}
+
+// capture executes the CRIU dump and post-dump captures (/dev/shm, rootfs diff).
+// Returns the CRIU dump duration for timing reporting.
+func (c *Checkpointer) capture(
+	criuOpts *criurpc.CriuOpts,
+	data *CheckpointMetadata,
+	state *containerState,
+	checkpointDir string,
+) (time.Duration, error) {
+	// Execute CRIU dump
+	criuDumpStart := time.Now()
+	criuClient := criu.MakeCriu()
+	if err := criuClient.Dump(criuOpts, nil); err != nil {
+		c.log.WithField("duration", time.Since(criuDumpStart)).Error("CRIU dump failed")
+		return 0, fmt.Errorf("CRIU dump failed: %w", err)
+	}
+	criuDumpDuration := time.Since(criuDumpStart)
+	c.log.WithField("duration", criuDumpDuration).Info("CRIU dump completed")
+
+	// Capture /dev/shm contents (must happen after dump for final process state)
+	if err := CaptureDevShm(state.PID, checkpointDir, c.log); err != nil {
+		c.log.WithError(err).Warn("Failed to capture /dev/shm contents")
+	}
+
+	// Capture rootfs diff and deleted files
+	CaptureRootfsState(state.UpperDir, checkpointDir, data, c.log)
+
+	return criuDumpDuration, nil
 }
