+++ deploy/chrek/pkg/checkpoint/namespaces.go
@@ -3,13 +3,16 @@ package checkpoint
 
 import (
 	"fmt"
-	"os"
-	"strings"
 
 	"golang.org/x/sys/unix"
-
-	"github.com/ai-dynamo/dynamo/deploy/chrek/pkg/config"
 )
+
+// NamespaceMetadata stores namespace information saved in checkpoint metadata.
+type NamespaceMetadata struct {
+	Type       string `yaml:"type"`       // net, pid, mnt, etc.
+	Inode      uint64 `yaml:"inode"`      // Namespace inode
+	IsExternal bool   `yaml:"isExternal"` // Whether namespace is external (shared)
+}
 
 // NamespaceType represents a Linux namespace type
 type NamespaceType string
@@ -28,13 +31,29 @@ const (
 type NamespaceInfo struct {
 	Type       NamespaceType
 	Inode      uint64
-	Path       string
 	IsExternal bool // Whether NS is external (shared with pause container)
+}
+
+// NewNamespaceMetadata constructs namespace metadata from introspected namespaces.
+func NewNamespaceMetadata(namespaces map[NamespaceType]*NamespaceInfo) []NamespaceMetadata {
+	if len(namespaces) == 0 {
+		return nil
+	}
+
+	result := make([]NamespaceMetadata, 0, len(namespaces))
+	for nsType, nsInfo := range namespaces {
+		result = append(result, NamespaceMetadata{
+			Type:       string(nsType),
+			Inode:      nsInfo.Inode,
+			IsExternal: nsInfo.IsExternal,
+		})
+	}
+	return result
 }
 
 // GetNamespaceInode returns the inode number for a namespace
 func GetNamespaceInode(pid int, nsType NamespaceType) (uint64, error) {
-	nsPath := fmt.Sprintf("%s/%d/ns/%s", config.HostProcPath, pid, nsType)
+	nsPath := fmt.Sprintf("%s/%d/ns/%s", HostProcPath, pid, nsType)
 	var stat unix.Stat_t
 	if err := unix.Stat(nsPath, &stat); err != nil {
 		return 0, fmt.Errorf("failed to stat namespace %s: %w", nsPath, err)
@@ -45,7 +64,7 @@ func GetNamespaceInode(pid int, nsType NamespaceType) (uint64, error) {
 
 // GetNamespaceInfo returns detailed namespace information
 func GetNamespaceInfo(pid int, nsType NamespaceType) (*NamespaceInfo, error) {
-	nsPath := fmt.Sprintf("%s/%d/ns/%s", config.HostProcPath, pid, nsType)
+	nsPath := fmt.Sprintf("%s/%d/ns/%s", HostProcPath, pid, nsType)
 
 	// Get inode
 	var stat unix.Stat_t
@@ -53,14 +72,8 @@ func GetNamespaceInfo(pid int, nsType NamespaceType) (*NamespaceInfo, error) {
 		return nil, fmt.Errorf("failed to stat namespace %s: %w", nsPath, err)
 	}
 
-	// Read the symlink to get the namespace identifier
-	link, err := os.Readlink(nsPath)
-	if err != nil {
-		return nil, fmt.Errorf("failed to readlink %s: %w", nsPath, err)
-	}
-
 	// Check if this is different from init's namespace (PID 1)
-	initNsPath := fmt.Sprintf("%s/1/ns/%s", config.HostProcPath, nsType)
+	initNsPath := fmt.Sprintf("%s/1/ns/%s", HostProcPath, nsType)
 	var initStat unix.Stat_t
 	isExternal := false
 	if err := unix.Stat(initNsPath, &initStat); err == nil {
@@ -71,7 +84,6 @@ func GetNamespaceInfo(pid int, nsType NamespaceType) (*NamespaceInfo, error) {
 	return &NamespaceInfo{
 		Type:       nsType,
 		Inode:      stat.Ino,
-		Path:       link,
 		IsExternal: isExternal,
 	}, nil
 }
@@ -101,51 +113,3 @@ func GetAllNamespaces(pid int) (map[NamespaceType]*NamespaceInfo, error) {
 	return namespaces, nil
 }
 
-// IsNetNamespaceExternal checks if the network namespace is external
-// (i.e., shared with the pause container in Kubernetes)
-func IsNetNamespaceExternal(pid int) (bool, uint64, error) {
-	info, err := GetNamespaceInfo(pid, NamespaceNet)
-	if err != nil {
-		return false, 0, err
-	}
-	return info.IsExternal, info.Inode, nil
-}
-
-// IsPIDNamespaceExternal checks if the PID namespace is external
-func IsPIDNamespaceExternal(pid int) (bool, uint64, error) {
-	info, err := GetNamespaceInfo(pid, NamespacePID)
-	if err != nil {
-		return false, 0, err
-	}
-	return info.IsExternal, info.Inode, nil
-}
-
-// OpenNamespaceFD opens a file descriptor to a namespace
-// The caller is responsible for closing the returned file
-func OpenNamespaceFD(pid int, nsType NamespaceType) (*os.File, error) {
-	nsPath := fmt.Sprintf("%s/%d/ns/%s", config.HostProcPath, pid, nsType)
-	return os.Open(nsPath)
-}
-
-// FormatExternalNamespace formats namespace info for CRIU's External option
-// Format: <type>[<inode>]:<key>
-func FormatExternalNamespace(nsType NamespaceType, inode uint64) string {
-	key := formatNamespaceKey(nsType)
-	return fmt.Sprintf("%s[%d]:%s", nsType, inode, key)
-}
-
-// formatNamespaceKey creates the CRIU key for external namespaces
-// Format: extRoot<Type>NS (e.g., extRootNetNS, extRootPidNS)
-func formatNamespaceKey(nsType NamespaceType) string {
-	// Capitalize first letter of namespace type
-	nsName := string(nsType)
-	if len(nsName) > 0 {
-		nsName = strings.ToUpper(nsName[:1]) + nsName[1:]
-	}
-	return "extRoot" + nsName + "NS"
-}
-
-// GetNamespaceKey returns the CRIU key for a namespace type
-func GetNamespaceKey(nsType NamespaceType) string {
-	return formatNamespaceKey(nsType)
-}
