+++ deploy/chrek/pkg/restore/config.go
@@ -1,10 +1,10 @@
-// restore.go defines the RestoreConfig struct for CRIU restore operations.
-// CRIU options come from the saved CheckpointData, not from this config.
+// config.go defines the RestoreConfig struct for CRIU restore operations.
+// CRIU options come from the saved CheckpointMetadata, not from this config.
 //
 // The restore-entrypoint runs in placeholder containers which do NOT mount the
 // ConfigMap. Static defaults are hardcoded here; per-pod dynamic values come
 // from environment variables injected by the operator.
-package config
+package restore
 
 import (
 	"context"
@@ -14,44 +14,47 @@ import (
 	"time"
 
 	"github.com/sirupsen/logrus"
+
+	"github.com/ai-dynamo/dynamo/deploy/chrek/pkg/checkpoint"
+)
+
+const (
+	// RestoreLogFilename is the CRIU restore log filename.
+	RestoreLogFilename = "restore.log"
+
+	// CRIULogDir is the directory where CRIU restore logs are copied for debugging.
+	CRIULogDir = "/checkpoints/restore-logs"
+
+	// RestoreTriggerPath is the default path to the trigger file for trigger-based restore.
+	RestoreTriggerPath = "/tmp/restore-trigger"
 )
 
 // RestoreConfig holds the configuration for the restore entrypoint.
-// CRIU options are NOT stored here - they come from the saved CheckpointData.
+// CRIU options are NOT stored here - they come from the saved CheckpointMetadata.
 type RestoreConfig struct {
 	// === Per-pod dynamic values (from operator-injected env vars) ===
 
 	// CheckpointPath is the base directory containing checkpoints.
-	// From DYN_CHECKPOINT_PATH env var (operator-injected).
 	CheckpointPath string
 
 	// CheckpointHash is the ID/hash of the checkpoint to restore.
-	// From DYN_CHECKPOINT_HASH env var (operator-injected).
 	CheckpointHash string
 
 	// CheckpointLocation is the full resolved path to the checkpoint directory.
-	// Computed as CheckpointPath + "/" + CheckpointHash for PVC storage.
-	// For future S3/OCI support, DYN_CHECKPOINT_LOCATION takes priority if set.
 	CheckpointLocation string
 
 	// SkipWaitForCheckpoint controls the entrypoint behavior.
-	// When true (operator sets SKIP_WAIT_FOR_CHECKPOINT=1): check once, restore
-	// if ready, otherwise exec ColdStartArgs (cold start). No polling.
-	// When false (default, standalone): poll/wait for checkpoint.done to appear.
 	SkipWaitForCheckpoint bool
 
-	// ColdStartArgs is the command+args to exec if no checkpoint is available
-	// and SkipWaitForCheckpoint is true. Populated from os.Args[1:].
+	// ColdStartArgs is the command+args to exec if no checkpoint is available.
 	ColdStartArgs []string
 
 	// Debug enables debug logging.
-	// From DEBUG env var.
 	Debug bool
 
 	// === Static defaults (hardcoded) ===
 
 	// RestoreMarkerFilePath is where restore-entrypoint writes a marker before CRIU restore.
-	// From DYN_RESTORE_MARKER_FILE env var.
 	RestoreMarkerFilePath string
 
 	// RestoreTrigger is the path to the trigger file that signals restore should start.
@@ -62,16 +65,24 @@ type RestoreConfig struct {
 	WaitTimeout time.Duration
 }
 
+// ConfigError represents a configuration validation error.
+type ConfigError struct {
+	Field   string
+	Message string
+}
+
+func (e *ConfigError) Error() string {
+	return fmt.Sprintf("config error: %s: %s", e.Field, e.Message)
+}
+
 // NewRestoreConfig creates a RestoreConfig with hardcoded defaults and
 // operator-injected environment variable values.
 func NewRestoreConfig(args []string) (*RestoreConfig, error) {
 	cfg := &RestoreConfig{
-		// Static defaults
 		RestoreTrigger: RestoreTriggerPath,
 		ColdStartArgs:  args,
 	}
 
-	// Per-pod dynamic values from operator-injected env vars
 	if v := os.Getenv("DYN_CHECKPOINT_PATH"); v != "" {
 		cfg.CheckpointPath = v
 	}
@@ -79,9 +90,6 @@ func NewRestoreConfig(args []string) (*RestoreConfig, error) {
 		cfg.CheckpointHash = v
 	}
 
-	// Compute checkpoint location.
-	// DYN_CHECKPOINT_LOCATION takes priority (future S3/OCI support).
-	// Otherwise construct from PATH + "/" + HASH for PVC storage.
 	if v := os.Getenv("DYN_CHECKPOINT_LOCATION"); v != "" {
 		cfg.CheckpointLocation = v
 	} else if cfg.CheckpointPath != "" && cfg.CheckpointHash != "" {
@@ -103,14 +111,10 @@ func NewRestoreConfig(args []string) (*RestoreConfig, error) {
 
 // ShouldRestore checks if a restore should be performed.
 // Returns the checkpoint path and true if restore should proceed.
-// IMPORTANT: We check for checkpoint.done marker (not just metadata.yaml or inventory.img) because
-// checkpoint.done is written LAST in the checkpoint process, after rootfs-diff.tar completes.
-// Order: metadata.yaml -> CRIU dump (*.img files) -> rootfs-diff.tar -> checkpoint.done
 func ShouldRestore(cfg *RestoreConfig, log *logrus.Entry) (string, bool) {
 	// Method 1: Checkpoint location is set and checkpoint is fully complete
 	if cfg.CheckpointLocation != "" {
-		// Check for checkpoint.done marker (written LAST after rootfs-diff.tar completes)
-		donePath := cfg.CheckpointLocation + "/" + CheckpointDoneFilename
+		donePath := cfg.CheckpointLocation + "/" + checkpoint.CheckpointDoneFilename
 
 		if _, err := os.Stat(donePath); err == nil {
 			log.WithField("path", cfg.CheckpointLocation).Info("Checkpoint found (checkpoint.done marker present)")
@@ -118,13 +122,12 @@ func ShouldRestore(cfg *RestoreConfig, log *logrus.Entry) (string, bool) {
 		}
 
 		// Fallback: check for metadata.yaml but warn about potential race condition
-		metadataPath := cfg.CheckpointLocation + "/" + CheckpointDataFilename
+		metadataPath := cfg.CheckpointLocation + "/" + checkpoint.CheckpointDataFilename
 		if _, err := os.Stat(metadataPath); err == nil {
 			log.WithFields(logrus.Fields{
 				"path":    cfg.CheckpointLocation,
 				"warning": "checkpoint.done marker not found, checkpoint may be incomplete",
 			}).Warn("Checkpoint metadata found but checkpoint.done missing - checkpoint may still be in progress")
-			// Don't return true here - wait for checkpoint.done
 		}
 	}
 
@@ -134,7 +137,7 @@ func ShouldRestore(cfg *RestoreConfig, log *logrus.Entry) (string, bool) {
 		if err == nil {
 			checkpointPath := strings.TrimSpace(string(data))
 			if checkpointPath != "" {
-				donePath := checkpointPath + "/" + CheckpointDoneFilename
+				donePath := checkpointPath + "/" + checkpoint.CheckpointDoneFilename
 				if _, err := os.Stat(donePath); err == nil {
 					log.WithField("path", checkpointPath).Info("Restore triggered via file (checkpoint.done marker present)")
 					return checkpointPath, true
