+++ deploy/chrek/pkg/restore/mounts.go
@@ -6,19 +6,14 @@ import (
 	criurpc "github.com/checkpoint-restore/go-criu/v7/rpc"
 	"google.golang.org/protobuf/proto"
 
+	"github.com/ai-dynamo/dynamo/deploy/chrek/pkg/checkpoint"
 	"github.com/ai-dynamo/dynamo/deploy/chrek/pkg/common"
-	"github.com/ai-dynamo/dynamo/deploy/chrek/pkg/config"
 )
-
-// Deprecated type alias for backwards compatibility
-type CheckpointMetadata = config.CheckpointData
 
 // GenerateExtMountMaps generates external mount mappings for CRIU restore.
 // It parses /proc/1/mountinfo (the restore container's mounts) and adds
-// mappings for all mount points plus masked/readonly/bind mount paths from checkpoint data.
-//
-// If data is nil or doesn't have OCI-derived paths, falls back to defaults.
-func GenerateExtMountMaps(data *config.CheckpointData) ([]*criurpc.ExtMountMap, error) {
+// mappings for all mount points plus masked/readonly/bind mount paths from checkpoint metadata.
+func GenerateExtMountMaps(data *checkpoint.CheckpointMetadata) ([]*criurpc.ExtMountMap, error) {
 	var maps []*criurpc.ExtMountMap
 	addedMounts := make(map[string]bool)
 
@@ -47,11 +42,11 @@ func GenerateExtMountMaps(data *config.CheckpointData) ([]*criurpc.ExtMountMap, 
 		addedMounts[m.Path] = true
 	}
 
-	// Use masked paths from checkpoint data (OCI spec derived)
+	// Use masked paths from checkpoint metadata (OCI spec derived)
 	// Fall back to defaults for backwards compatibility
 	maskedPaths := common.DefaultMaskedPaths()
-	if data != nil && len(data.MaskedPaths) > 0 {
-		maskedPaths = data.MaskedPaths
+	if data != nil && len(data.Filesystem.MaskedPaths) > 0 {
+		maskedPaths = data.Filesystem.MaskedPaths
 	}
 
 	for _, path := range maskedPaths {
@@ -65,9 +60,9 @@ func GenerateExtMountMaps(data *config.CheckpointData) ([]*criurpc.ExtMountMap, 
 		addedMounts[path] = true
 	}
 
-	// Also add readonly paths from checkpoint data if available
+	// Also add readonly paths from checkpoint metadata if available
 	if data != nil {
-		for _, path := range data.ReadonlyPaths {
+		for _, path := range data.Filesystem.ReadonlyPaths {
 			if addedMounts[path] {
 				continue
 			}
@@ -78,9 +73,8 @@ func GenerateExtMountMaps(data *config.CheckpointData) ([]*criurpc.ExtMountMap, 
 			addedMounts[path] = true
 		}
 
-		// Add bind mount destinations from checkpoint data
-		// These are critical for CRIU to properly map checkpoint mounts to restore mounts
-		for _, path := range data.BindMountDests {
+		// Add bind mount destinations from checkpoint metadata
+		for _, path := range data.Filesystem.BindMountDests {
 			if addedMounts[path] {
 				continue
 			}
@@ -91,8 +85,7 @@ func GenerateExtMountMaps(data *config.CheckpointData) ([]*criurpc.ExtMountMap, 
 			addedMounts[path] = true
 		}
 
-		// Also add container paths from mount data
-		// This ensures all mounts from the checkpoint have mappings
+		// Also add container paths from mount metadata
 		for _, mount := range data.Mounts {
 			if addedMounts[mount.ContainerPath] {
 				continue
