+++ deploy/chrek/pkg/restore/shm.go
@@ -9,22 +9,14 @@ import (
 
     "github.com/sirupsen/logrus"
 
-    "github.com/ai-dynamo/dynamo/deploy/chrek/pkg/config"
+    "github.com/ai-dynamo/dynamo/deploy/chrek/pkg/checkpoint"
 )
 
 // RestoreDevShm restores files from the checkpoint's dev-shm directory to /dev/shm.
 // This must be called BEFORE CRIU restore so that the shared memory files exist
 // when CRIU tries to restore file descriptors pointing to them.
-//
-// /dev/shm may contain various types of files including:
-//   - POSIX semaphores (sem.*)
-//   - PSM/shared memory files used by communication libraries
-//   - Other tmpfs-backed shared memory regions
-//
-// The kernel state for these files won't be perfectly restored, but the files
-// will exist for CRIU to restore file descriptors and for cleanup operations.
 func RestoreDevShm(checkpointPath string, log *logrus.Entry) error {
-    srcDir := filepath.Join(checkpointPath, config.DevShmDirName)
+    srcDir := filepath.Join(checkpointPath, checkpoint.DevShmDirName)
 
     // Check if dev-shm directory exists in checkpoint
     entries, err := os.ReadDir(srcDir)
